requirements.txt
phidata
python-dotenv
groq
pycountry
googlesearch-python
yfinance
fastapi
uvicorn
openai